Streamline the app setup process
This adds a generator to more easily create a credentials file with valid VAPID configuration (necessary for sending web push notifications). This also updates the README to more accurately describe the setup process. The setup script was removed due to a potential race condition with postgres initialization. The docker-compose configuration sets the EDITOR=vim environment variable to simplify the process of editing the credentials.